cli refactor no folders
