+++ openfecli/commands/plan_rbfe_network.py
@@ -1,28 +1,28 @@
 # This code is part of OpenFE and is licensed under the MIT license.
 # For details, see https://github.com/OpenFreeEnergy/openfe
 
-
 """
-Here I want to build the cmd tool for easy campaigner with RHFE. The output should be runnable with quickrun directly!
-    So user experience would be:
-        easy_campaign -l sdf_dir -p receptor.pdb -> Alchem network
-        quickrun -i alchem_network
+
 
 """
 
 import click
 from typing import List
-
 from openfecli.utils import write
 from openfecli import OFECommandPlugin
-from openfecli.parameters import MOL_DIR, MAPPER, OUTPUT_DIR
-from openfecli.commands.alchemical_network_planners.utils import plan_alchemical_network_output
+from openfecli.parameters import MOL_DIR, PROTEIN, MAPPER, OUTPUT_DIR
+from openfecli.commands.plan_alchemical_networks_utils import plan_alchemical_network_output
 
 
-def plan_rhfe_network_main(
-    mapper, mapping_scorer, ligand_network_planner, small_molecules, solvent
+def plan_rbfe_network_main(
+    mapper,
+    mapping_scorer,
+    ligand_network_planner,
+    small_molecules,
+    solvent,
+    protein,
 ):
-    """Utility method to plan a relative hydration free energy network.
+    """Utiiity method to plan a relative binding free energy network.
 
     Parameters
     ----------
@@ -33,50 +33,55 @@ def plan_rhfe_network_main(
     ligand_network_planner : Callable
         function building the network from the ligands, mappers and mapping_scorer
     small_molecules : Iterable[SmallMoleculeComponent]
-        molecules of the system
+        ligands of the system
     solvent : SolventComponent
         Solvent component used for solvation
+    protein : ProteinComponent
+        protein component for complex simulations, to which the ligands are bound
 
     Returns
     -------
     AlchemicalNetwork
         Alchemical network with protocol for executing simulations.
     """
+
     from openfe.setup.alchemical_network_planner.relative_alchemical_network_planner import (
-        RHFEAlchemicalNetworkPlanner,
+        RBFEAlchemicalNetworkPlanner,
     )
 
-    network_planner = RHFEAlchemicalNetworkPlanner(
+    network_planner = RBFEAlchemicalNetworkPlanner(
         mappers=[mapper],
         mapping_scorer=mapping_scorer,
         ligand_network_planner=ligand_network_planner,
     )
     alchemical_network = network_planner(
-        ligands=small_molecules, solvent=solvent
+        ligands=small_molecules, solvent=solvent, protein=protein
     )
 
     return alchemical_network
 
 
 @click.command(
-    "plan-rhfe-network",
-    short_help="Run a planning session for relative hydration free energies, "
-               "saved in a dir with multiple JSON files (future: will be one JSON file)",
+    "plan-rbfe-campaign",
+    short_help="Run a planning session for relative binding free energies, saved in a dir with multiple JSON files (future: will be one JSON file)",
 )
 @MOL_DIR.parameter(
     required=True, help=MOL_DIR.kwargs["help"] + " Any number of sdf paths."
+)
+@PROTEIN.parameter(
+    multiple=False, required=True, default=None, help=PROTEIN.kwargs["help"]
 )
 @OUTPUT_DIR.parameter(
     help=OUTPUT_DIR.kwargs["help"] + " Defaults to `./alchemicalNetwork`.",
     default="alchemicalNetwork",
 )
 @MAPPER.parameter(required=False, default="LomapAtomMapper")
-def plan_rhfe_network(mol_dir: List[str], output_dir: str, mapper: str):
+def plan_rbfe_network(
+    mol_dir: List[str], protein: str, output_dir: str, mapper: str
+):
     """
-    this command line tool allows relative hydration free energy calculations
-    to be planned and returns an alchemical network as a directory.
+    this command line tool allows relative binding free energy calculations to be planned and returns an alchemical network as a directory.
     """
-
     from gufe import SolventComponent
     from openfe.setup.atom_mapping.lomap_scorers import (
         default_lomap_score,
@@ -85,11 +90,11 @@ def plan_rhfe_network(mol_dir: List[str], output_dir: str, mapper: str):
         generate_minimal_spanning_network,
     )
 
-    write("RHFE-NETWORK PLANNER")
+    # INPUT
+    write("RBFE-NETWORK PLANNER")
     write("______________________")
     write("")
 
-    # INPUT
     write("Parsing in Files: ")
     write("\tGot input: ")
 
@@ -98,6 +103,9 @@ def plan_rhfe_network(mol_dir: List[str], output_dir: str, mapper: str):
         "\t\tSmall Molecules: "
         + " ".join([str(sm) for sm in small_molecules])
     )
+
+    protein = PROTEIN.get(protein)
+    write("\t\tProtein: " + str(protein))
 
     solvent = SolventComponent()
     write("\t\tSolvent: " + str(solvent))
@@ -111,26 +119,27 @@ def plan_rhfe_network(mol_dir: List[str], output_dir: str, mapper: str):
     mapping_scorer = default_lomap_score
     write("\tMapping Scorer: " + str(mapping_scorer))
 
-    # TODO: write nice parameter
+    # TODO:  write nice parameter
     ligand_network_planner = generate_minimal_spanning_network
     write("\tNetworker: " + str(ligand_network_planner))
     write("")
 
     # DO
-    write("Planning RHFE-Campaign:")
-    alchemical_network = plan_rhfe_network_main(
+    write("Planning RBFE-Campaign:")
+    alchemical_network = plan_rbfe_network_main(
         mapper=mapper_obj,
         mapping_scorer=mapping_scorer,
         ligand_network_planner=ligand_network_planner,
         small_molecules=small_molecules,
         solvent=solvent,
+        protein=protein,
     )
     write("\tDone")
     write("")
 
     # OUTPUT
     write("Output:")
-    write("\tSaving to: " + output_dir)
+    write("\tSaving to: " + str(output_dir))
     plan_alchemical_network_output(
         alchemical_network=alchemical_network,
         folder_path=OUTPUT_DIR.get(output_dir),
@@ -138,5 +147,5 @@ def plan_rhfe_network(mol_dir: List[str], output_dir: str, mapper: str):
 
 
 PLUGIN = OFECommandPlugin(
-    command=plan_rhfe_network, section="Setup", requires_ofe=(0, 3)
+    command=plan_rbfe_network, section="Setup", requires_ofe=(0, 3)
 )
